requirements.txt
aiofiles==23.2.1
altair==5.1.1
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.4.0
anndata==0.8.0
annotated-types==0.5.0
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-lru==2.0.4
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.5
blosc2==2.2.9
bokeh==3.1.1
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==2.2.0
cmake==3.25.2
colorama==0.4.6
colorcet==3.0.1
comm==0.1.4
conjure-python-client==2.4.0
contextvars==2.4
contourpy==1.0.6
cryptography==41.0.4
cycler==0.11.0
Cython==0.29.33
dask==2022.12.1
datashader==0.15.2
datashape==0.5.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dict-toolbox==3.1.2
dill==0.3.5.1
dlhub-sdk==2.0.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fair-research-login==0.3.1
fastjsonschema==2.16.2
fcsparser==0.2.4
fonttools==4.42.1
foundry-ml==0.5.1
fqdn==1.5.1
fsspec==2022.11.0
funcx==2.0.1
gitdb==4.0.10
GitPython==3.1.37
globus-compute-common==0.2.0
globus-compute-sdk==2.3.2
globus-nexus-client==0.4.1
globus-sdk==3.21.0
h11==0.14.0
h5py==3.9.0
holoviews==1.15.3
htmltools==0.3.0
idna==3.4
igraph==0.10.6
imageio==2.31.1
immutables==0.19
importlib-metadata==6.0.0
ipykernel==6.25.1
ipytablewidgets==0.3.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jax==0.4.14
jaxlib==0.4.14
jaxopt==0.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json2table==1.1.5
json5==0.9.14
jsonpointer==2.3
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.25.0
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lazy_loader==0.2
leidenalg==0.10.1
linkify-it-py==2.0.0
llvmlite==0.40.0
locket==1.0.0
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdf-connect-client==0.4.1
mdf-forge==0.8.0
mdf-toolbox==0.6.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mellon==1.3.1
mistune==2.0.4
ml-dtypes==0.2.0
msgpack==1.0.4
multipledispatch==0.6.0
nameparser==1.1.2
natsort==8.4.0
nbclassic==0.4.8
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
ndindex==1.7
nest-asyncio==1.5.6
networkx==3.1
notebook==7.0.3
notebook_shim==0.2.2
numba==0.57.0
numexpr==2.8.4
numpy==1.24.4
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.4.0
packaging==23.0
palantir==1.3.0
palantir-sdk==0.12.0
pandas==1.5.3
pandocfilters==1.5.0
panel==1.1.1
param==1.12.3
parso==0.8.3
partd==1.3.0
patsy==0.5.3
PhenoGraph==1.5.7
pickleshare==0.7.5
pika==1.3.1
Pillow==10.0.1
platformdirs==2.6.2
pop==26.0.4
pop-config==12.0.2
pop-loop==1.1.0
portpicker==1.5.2
progressbar2==4.2.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pycparser==2.21
pyct==0.4.8
pydantic==1.10.12
pydantic_core==2.6.3
pydeck==0.8.0
pygam==0.9.0
pygame==2.5.1
PyGithub==1.59.1
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
PyNaCl==1.5.0
pynndescent==0.5.8
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
python-multipart==0.0.5
python-utils==3.7.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.4.0
referencing==0.30.2
rend==6.5.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.0
scanpy==1.9.1
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.13.0
selenium==4.11.2
semver==2.13.0
Send2Trash==1.8.2
session-info==1.0.0
shiny==0.5.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.31.1
statsmodels==0.13.5
stdlib-list==0.8.0
streamlit==1.27.2
streamlit-javascript==0.1.5
tables==3.9.1
tblib==3.0.0
tenacity==8.2.3
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.1.0
tifffile==2023.8.25
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.8.1
traittypes==0.2.1
trio==0.22.2
trio-websocket==0.9.2
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.2
uc-micro-py==1.0.1
umap==0.1.1
umap-learn==0.5.4
uri-template==1.2.0
urllib3==2.0.7
urllib3-secure-extra==0.1.0
uvicorn==0.23.2
validators==0.20.0
vega==4.0.0
vega-datasets==0.9.0
vl-convert-python==0.13.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.3
widgetsnbextension==4.0.8
wrapt==1.15.0
wsproto==1.2.0
xarray==2023.10.1
xyzservices==2023.7.0
zipp==3.17.0
